workshops/enterprise-data-agent-harness-workshop/app/backend/requirements.txt
flask>=3.0
flask-socketio>=5.3
flask-cors>=4.0
python-socketio>=5.10
eventlet>=0.35

oracledb>=2.4
oracleagentmemory>=26.4
openai>=1.50

numpy<2.0
python-dotenv>=1.0
tavily-python>=0.5
